+++ src/App.css
@@ -26,6 +26,10 @@
   color: rgb(112, 76, 182);
 }
 
+.center {
+  text-align: center;
+}
+
 @keyframes App-logo-float {
   0% {
     transform: translateY(0);
