+++ src/Component/Dash.js
@@ -1,28 +1,12 @@
-import React, { useState } from 'react';
-import clsx from 'clsx';
-import { GoogleLogout } from 'react-google-login';
+import React, { useCallback, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Button } from '@material-ui/core';
 import Grid from '@material-ui/core/Grid';
 import { makeStyles } from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
-import Drawer from '@material-ui/core/Drawer';
-import AddBoxIcon from '@mui/icons-material/AddBox';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import List from '@material-ui/core/List';
 import Container from '@material-ui/core/Container';
-import Typography from '@material-ui/core/Typography';
-import Divider from '@material-ui/core/Divider';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
 import Box from '@mui/material/Box';
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
 import Paper from '@mui/material/Paper';
-
 import TextField from '@material-ui/core/TextField';
 import AdapterDateFns from '@mui/lab/AdapterDateFns';
 import LocalizationProvider from '@mui/lab/LocalizationProvider';
@@ -31,7 +15,8 @@ import DatePicker from '@mui/lab/DatePicker';
 import CurrentAddedLeaves from './CurrentAddedLeaves';
 import PerticularLeaves from './PerticularLeaves';
 import Title from './Title';
-import AddLeaveDialog from './AddLeaveDialog';
+import AppBarDrawer from './AppBarDrawer';
+import '../App.css';
 
 const drawerWidth = 240;
 const useStyles = makeStyles((theme) => ({
@@ -64,7 +49,7 @@ const useStyles = makeStyles((theme) => ({
     }),
   },
   menuButton: {
-    marginRight: 36,
+    marginRight: '1rem',
   },
   menuButtonHidden: {
     display: 'none',
@@ -109,7 +94,7 @@ const useStyles = makeStyles((theme) => ({
     flexDirection: 'column',
   },
   fixedHeight: {
-    height: 240,
+    height: '240rem',
   },
 }));
 
@@ -130,138 +115,17 @@ function Dash() {
     navigate('/');
   };
 
-  const [openDrawer, setOpenDrawer] = useState(false);
-  const handleDrawerOpen = () => {
-    setOpenDrawer(true);
-  };
-  const handleDrawerClose = () => {
-    setOpenDrawer(false);
-  };
-
-  const [isAddLeaves, setAddLeaves] = useState(false);
   const [pLeave, setPLeave] = useState(false);
   const [date, setDate] = useState('');
-
-  const openDiag = () => {
-    setAddLeaves(true);
-  };
-
-  const closeDiag = () => {
-    setAddLeaves(false);
-  };
 
   const handleSubmit = () => {
     setPLeave(true);
   };
 
-  // declare a new state variable for modal open
-  const [open, setOpen] = useState(false);
-
-  // function to handle modal open
-  const handleOpen = () => {
-    setOpen(true);
-  };
-
-  // function to handle modal close
-  const handleClose = () => {
-    setOpen(false);
-  };
-
   return (
     <div className={classes.root}>
       <CssBaseline />
-      <AppBar
-        style={{ background: '#15317E' }}
-        position="absolute"
-        className={clsx(classes.appBar, openDrawer && classes.appBarShift)}
-      >
-        <Toolbar className={classes.toolbar}>
-          <IconButton
-            edge="start"
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            className={clsx(
-              classes.menuButton,
-              openDrawer && classes.menuButtonHidden,
-            )}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Typography
-            component="h1"
-            variant="h6"
-            color="inherit"
-            noWrap
-            className={classes.title}
-          >
-            Leave Tracker
-          </Typography>
-          <Typography
-            component="h3"
-            variant="h6"
-            color="inherit"
-            noWrap
-            className={classes.title}
-          >
-            Hi,{' '}
-            {localStorage.getItem('first_name') +
-              localStorage.getItem('last_name')}
-          </Typography>
-          <Button variant="contained" color="sucess" onClick={handleOpen}>
-            Add Leaves
-          </Button>
-          <AddLeaveDialog open={open} handleClose={handleClose} />
-          <GoogleLogout
-            clientId={process.env.REACT_APP_CLIENT_KEY}
-            render={(renderProps) => (
-              <Box sx={{ mx: 2 }}>
-                <Button
-                  variant="contained"
-                  color="primary"
-                  style={{ background: '#D2042D' }}
-                  onClick={renderProps.onClick}
-                >
-                  Logout
-                </Button>
-              </Box>
-            )}
-            buttonText="Logout"
-            icon={false}
-            onLogoutSuccess={logout}
-          />
-        </Toolbar>
-      </AppBar>
-      <Drawer
-        variant="permanent"
-        classes={{
-          paper: clsx(
-            classes.drawerPaper,
-            !openDrawer && classes.drawerPaperClose,
-          ),
-        }}
-        open={openDrawer}
-      >
-        <div className={classes.toolbarIcon}>
-          <IconButton onClick={handleDrawerClose}>
-            <ChevronLeftIcon />
-          </IconButton>
-        </div>
-        <Divider />
-        <List>
-          <div>
-            <ListItem button onClick={openDiag}>
-              {isAddLeaves && (
-                <AddLeaveDialog open={isAddLeaves} handleClose={closeDiag} />
-              )}
-              <ListItemIcon>
-                <AddBoxIcon />
-              </ListItemIcon>
-              <ListItemText primary="Add Leaves" />
-            </ListItem>
-          </div>
-        </List>
-      </Drawer>
+      <AppBarDrawer />
       <main className={classes.content}>
         <div className={classes.appBarSpacer} />
         <Container maxWidth="lg" className={classes.container}>
